requirement.txt
streamlit==1.35.0
pandas==2.2.2
pyyaml==6.0.1
promptfoo==0.115.2
matplotlib==3.8.4
seaborn==0.13.2